# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Balance API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local"
        }
    ],
    "paths": {
        "/api/balance/{userId}": {
            "get": {
                "tags": [
                    "Finance"
                ],
                "summary": "Текущий баланс пользователя",
                "operationId": "1c3474b7424c763ec56a8de85175d15c",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ок",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "balance": {
                                            "type": "string",
                                            "example": "150.00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}